+++ src/main/java/kitchenpos/menu/dto/MenuResponse.java
@@ -1,9 +1,9 @@
-package kitchenpos.dto;
+package kitchenpos.menu.dto;
 
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 
 public class MenuResponse {
