+++ src/test/java/kitchenpos/menu/MenuGroupAcceptanceTest.java
@@ -8,7 +8,7 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
 import kitchenpos.AcceptanceTest;
-import kitchenpos.dto.MenuGroupResponse;
+import kitchenpos.menu.dto.MenuGroupResponse;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
